requirements.txt
Django==1.3.1
PIL==1.1.7
django-admin-tools==0.4.0
django-annoying==0.7.6
django-appconf==0.4.1
django-compressor==1.1.1
django-filebrowser==3.3.0
django-grappelli==2.3.5
django-mailer==0.1.0
django-notification==0.1.5
django-profiles==0.2
django-registration==0.7
django-robots==0.8.0
django-rosetta==0.6.2
django-transmeta==0.6.2
django-urli18n==0.1
feedparser==5.0.1
lxml==2.3.1
mutagen==1.20
wsgiref==0.1.2
yolk==0.4.1
